+++ src/main/java/de/codescape/jira/plugins/scrumpoker/action/ShowSessionsAction.java
@@ -8,8 +8,8 @@ import com.atlassian.jira.security.PermissionManager;
 import com.atlassian.jira.user.ApplicationUser;
 import com.atlassian.jira.user.util.UserManager;
 import com.atlassian.jira.web.action.JiraWebActionSupport;
-import de.codescape.jira.plugins.scrumpoker.data.ScrumPokerStorage;
 import de.codescape.jira.plugins.scrumpoker.model.ScrumPokerSession;
+import de.codescape.jira.plugins.scrumpoker.persistence.ScrumPokerStorage;
 
 import java.util.List;
 import java.util.stream.Collectors;