+++ changelog.md
@@ -5,6 +5,7 @@ This changelog helps developers and users to keep track of new features, fixes a
 ## Unreleased
 
 * feat: add post update and post install urls to improve setup process
+* refactor: move all persistence related classes into persistence package
 
 ## 2.3.0 - 2018-04-11